extracting helper function for server-side error handling + changing error color for white and mix inputFields
