+++ App/Components/PlanGame/PlanGameApiCall/utils.js
@@ -1,4 +1,4 @@
-import isString from 'lodash/isString';
+import { curateErrors as ce } from '../../../utils';
 
 /**
  * error = { email: ['This field must be unique.'] }
@@ -30,35 +30,6 @@ const curateErrorMsg = (errorMsg) => {
   return errorMsg;
 };
 
-/**
- * errors = {
- *  email: ['This field must be unique.'],
- *  password1: ['This password is too short. It must contain at least 8 characters.']
- * }
- * OR
- * errors = ['Invalid Spot. Spot being assigned doesnt have the already associated sport']
-*/
-// TODO: make it DRY
-const curateErrors = (errors) => {
-  if (isString(errors)) {
-    return { [curateFieldName(null)]: [errors] }; // curatedErrors
-  }
-
-  if (Array.isArray(errors)) {
-    return { [curateFieldName(null)]: errors }; // curatedErrors
-  }
-
-  // In case errors is an object
-  const keys = Object.keys(errors);
-  const curatedErrors = {};
-
-  keys.forEach((key) => {
-    const arrayError = errors[key];
-    const curatedArray = arrayError.map(errorMsg => (curateErrorMsg(errorMsg)));
-    curatedErrors[curateFieldName(key)] = curatedArray;
-  });
-
-  return curatedErrors;
-};
+const curateErrors = ce(curateFieldName, curateErrorMsg);
 
 export default curateErrors;
